refactor: IntegerUtils 삭제
