+++ src/main/java/me/daeho/utils/StringUtils.java
@@ -1,0 +1,8 @@
+package me.daeho.utils;
+
+public class StringUtils {
+    private static final String BLANK = "";
+    public static boolean isEmpty(String data) {
+        return data == null || BLANK.equals(data.trim());
+    }
+}
